Database Indexing
The goal of this project is to implement a program that parses an ASCII
text-based CSV data file and then transforms that same information to a
memory-efficient format and writes it to a binary data file. The new
binary data file shall also have multiple index files for efficient
record retrieval of values from fields other than the primary key.